refactor hospitality hub items fetch
